Introduce SQL based statistics service.
This introduces StatisticService, a SQL based service providing
statistics on how many km a user has been bicking etc.

It replaces the old, excessive Java 8 stream API usage. Back in 2014, it
seemed cool, doing it all on the client, but the right place to compute
is, is the database.

Even H2 supports all the analytic functions (as window functions) needed,
to compute the runnign sums from the recorded absolute milages.

Together with a handful of common table expressions, everything boils
down to 5 queries.

Those have been tested by creating a new empty database, inserting the
old test data, and then querying it.

With that setup, I want to be in complete control again over my schema,
so I also removed Hibernates automatic schema creation and replaced it
with Flyway and SQL scripts.

A lot of tests have been simplified in the process.