Disable "open session in view".
This commit makes the `milages` association lazy (the default) and also disables the open session in view interceptor.

By storing the last recorded milage on the bike entity itself, the need for eagerly accessing the collection goes away.

The commit also removes `@Transactional` from the bikes controller while introducing a dedicated service for manipulating bikes.
The service takes care of the transaction handling from there on and reads the milages in a controlled way when manipulating them.

One fresh wound is the test of the biker controller. It loads the service into the test slice now and passes the mocked repository.
This shall be changed at some later point.